+++ src/canvas/components/mod.rs
@@ -3,5 +3,5 @@
 pub mod data_table;
 pub mod pipe_gauge;
 pub mod scroll_bar;
-pub mod time_graph;
+pub mod timeseries;
 pub mod widget_carousel;
