--- src/canvas/components/time_graph/variants/auto_y_axis.rs
@@ -1,2 +1,0 @@
-//! A variant of a [`crate::canvas::components::time_graph::TimeGraph`] that
-//! automatically adjusts the y-axis based on the data provided.
