+++ src/canvas/components/timeseries.rs
