+++ src/canvas/components/timeseries/variants.rs
