+++ src/canvas/components/timeseries/variants/auto_y_axis.rs
@@ -1,0 +1,2 @@
+//! A variant of a [`crate::canvas::components::timeseries::TimeGraph`] that
+//! automatically adjusts the y-axis based on the data provided.
