+++ src/canvas/widgets/mem_graph.rs
@@ -10,7 +10,7 @@ use crate::{
     app::{App, data::Values},
     canvas::{
         Painter,
-        components::time_graph::{GraphData, LegendConstraints, PercentTimeGraph},
+        components::timeseries::{GraphData, LegendConstraints, PercentTimeGraph},
         drawing_utils::should_hide_x_label,
     },
     collection::memory::MemData,
