+++ src/canvas/widgets/temperature_graph.rs
@@ -8,7 +8,7 @@ use crate::{
     app::{App, AppConfigFields},
     canvas::{
         Painter,
-        components::time_graph::{
+        components::timeseries::{
             AxisBound, ChartScaling, GraphData, LegendConstraints, TimeGraph,
         },
         drawing_utils::should_hide_x_label,
