+++ src/widgets/common/mod.rs
@@ -1,0 +1,5 @@
+//! Common widget code.
+
+mod timeseries;
+
+pub use timeseries::*;
